+++ ididi/_itypes.py
@@ -53,6 +53,8 @@ P8 = TypeVar("P8")
 P9 = TypeVar("P9")
 
 
+
+
 # Factory with many type params
 class TDecor(Protocol):
 
@@ -130,10 +132,10 @@ class NodeConfig:
 
 
 class GraphConfig:
-    __slots__ = "static_resolve"
+    __slots__ = "self_inject"
 
-    def __init__(self, *, static_resolve: bool = True):
-        self.static_resolve = static_resolve
+    def __init__(self, *, self_inject: bool = True):
+        self.self_inject = self_inject
 
 
 @runtime_checkable