+++ tests/features/test_dg_self_inject.py
@@ -1,4 +1,16 @@
-from ididi import DependencyGraph
+from ididi import DependencyGraph, Resource
+# from ididi._itypes import Resource
+
+
+class DataBase:
+    def __enter__(self): ...
+
+    def __exit__(self, *args): ...
+
+
+def get_db() -> Resource[DataBase]:
+    db = DataBase()
+    yield db
 
 
 class UserService: ...
@@ -11,3 +23,20 @@ def get_service(dg: DependencyGraph) -> UserService:
 def test_dg_self_inject():
     dg = DependencyGraph()
     assert dg is dg.resolve(get_service)
+
+
+def test_dg_register_dependent():
+    dg = DependencyGraph()
+    service = UserService()
+
+    dg.register_dependent(service, UserService)
+    assert dg.resolve(UserService) is service
+
+
+def test_dg_register_resource_dependent():
+    dg = DependencyGraph()
+    db = DataBase()
+    dg.node(get_db)
+    with dg.scope() as scope:
+        scope.register_dependent(db, DataBase)
+        assert scope.resolve(DataBase) is db