+++ tests/test_benchmark.py
@@ -12,7 +12,12 @@ ROUNDS = 10
 
 @pytest.fixture
 def dependents() -> ty.Sequence[type]:
-    deps = CLASSES.values()
+    # interesting
+    # deps = CLASSES.values() would let dg
+    # be substantially (3times) faster than [d for d in  ClASSES.values()]
+    # return CLASSES.values()
+
+    deps = [d for d in CLASSES.values()][:100]
     return deps
 
 
@@ -24,15 +29,15 @@ def dg() -> DependencyGraph:
 @pytest.mark.benchmark
 def test_register(dg: DependencyGraph, dependents: ty.Sequence[type]):
     pre = perf_counter()
-    for c in dependents:
-        dg.node(c)
+    for dep in dependents:
+        dg.node(dep)
     aft = perf_counter()
     cost = round(aft - pre, 6)
     print(f"\n{cost} seoncds to register {len(dependents)} classes")
 
 
 @pytest.mark.benchmark
-def test_static_resolve(dg: DependencyGraph, dependents: ty.Sequence[type]):
+def test_static_resolve(dg: DependencyGraph):
     pre = perf_counter()
     dg.static_resolve_all()
     aft = perf_counter()
@@ -71,4 +76,12 @@ def test_resolve_instances(dg: DependencyGraph, dependents: list[type]):
 0.019022 seoncds to register 122 classes
 0.00338 seoncds to statically resolve 122 classes
 0.001264 seoncds to resolve 122 instances
+"""
+
+"""
+1.1.0
+
+0.020025 seoncds to register 122 classes
+0.003467 seoncds to statically resolve 122 classes
+0.001032 seoncds to resolve 122 instances
 """